python/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
attrs==24.2.0
awsgi==0.0.5
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
dataclasses-json==0.6.7
deepgram-sdk==2.12.0
deprecation==2.1.0
dtw-python==1.5.3
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
marshmallow==3.23.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
propcache==0.2.0
regex==2024.9.11
requests==2.32.3
scipy==1.14.1
setuptools==75.2.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.0
torchaudio==2.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvloop==0.21.0
websockets==13.1
Werkzeug==3.0.4
whisper-timestamped==1.15.4
yarl==1.16.0